Moved Shape components from tetrisshapes.dart file into a separate file and reformatted imports
