+++ lib/dtetrisgame.dart
@@ -2,9 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 
-import 'package:tetris_game/tetrisshapes.dart';
-
-import 'package:tetris_game/dtetrisboard.dart';
+import 'components/shapes.dart';
 
 class Tetromino {
   final TetrimonosShape shape;
