+++ lib/tetrisgame1.dart
@@ -3,6 +3,9 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:tetris_game/tetrisshapes.dart';
 
+import 'components/shapes.dart';
+import 'components/tetris_block.dart';
+
 // void main() {
 //   runApp(const AppTetris());
 // }
